requirements.txt
annotated-types==0.7.0
certifi==2024.8.30
instabase-aihub==0.5.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
typing_extensions==4.12.2
urllib3==2.0.7
